requirements.txt
requests>=2.25.0
Pillow>=8.0.0
torch>=1.9.0
numpy>=1.20.0
python-osc>=1.8.0
